requirements.txt
torch==2.5.1
torchvision==0.20.1
gradio==5.23.1
pillow==10.4.0
numpy==2.2.2
opencv-python==4.10.0
matplotlib==3.10.0
wandb==0.19.6
tqdm==4.67.1
